+++ packages/router/CHANGELOG.md
@@ -1,5 +1,15 @@
 # `@remix-run/router`
 
+## 1.0.5
+
+### Patch Changes
+
+- Fix requests sent to revalidating loaders so they reflect a `GET` request ([#9680](https://github.com/remix-run/react-router/pull/9680))
+- Remove `instanceof Response` checks in favor of `isResponse` ([#9690](https://github.com/remix-run/react-router/pull/9690))
+- Fix `URL` creation in Cloudflare Pages or other non-browser-environments ([#9682](https://github.com/remix-run/react-router/pull/9682), [#9689](https://github.com/remix-run/react-router/pull/9689))
+- Add `requestContext` support to static handler `query`/`queryRoute` ([#9696](https://github.com/remix-run/react-router/pull/9696))
+  - Note that the unstable API of `queryRoute(path, routeId)` has been changed to `queryRoute(path, { routeId, requestContext })`
+
 ## 1.0.4
 
 ### Patch Changes
@@ -46,7 +56,7 @@ For an overview of the features provided by `react-router`, we recommend you go 
 
 For an overview of the features provided by `@remix-run/router`, please check out the [`README`][remix-router-readme].
 
-[rr-docs]: https://reactrouter.com/
-[rr-feature-overview]: https://reactrouter.com/en/6.4.0/start/overview
-[rr-tutorial]: https://reactrouter.com/en/6.4.0/start/tutorial
+[rr-docs]: https://reactrouter.com
+[rr-feature-overview]: https://reactrouter.com/start/overview
+[rr-tutorial]: https://reactrouter.com/start/tutorial
 [remix-router-readme]: https://github.com/remix-run/react-router/blob/main/packages/router/README.md